An error should be issued if a surface element does not have an id
It is a good practice to use classes to identify every element in the
page, even though it is supposed to exist only once.
Since senna requires an id for a surface instead of a class, it should
issue an error when a developer starts using senna without specifying
and id attribute.