Fixes test for repositiong scroll for hash anchors
The only reason this test was passing on some browsers (like Chrome) is because the native handling for scroll position without hashbangs happens asynchronously. The correct value of the scroll position at the end of the test was actually [0,0], not [1000, 1000], since the test was going back to the first page that had no anchors. Browsers without native handling were being handled by senna.js itself, and so had the correct value, and were failing because the test was wrong.

Since the idea of the test was to test restoring the scroll position when going back to paths with anchors, I moved the check to be before the last "history.back()" call, andthe test is now passing for all browsers.