fix: check for TypeError when making requests
in https://github.com/liferay/senna.js/pull/318/commits/2519577abcb8bc898a5af6bbbe908fbab59eae7d
a new constant was added to the `src/errors/errors.js` class to check
for "CORS" specific errors with fetch. The problem of this is that
each browser can issue a different error message, and it is actually
the case:
- Chrome uses `Failed to fetch`
- Firefox uses `NetworkError when attempting to fetch resource.`
- Safari uses `Failed to load resource: Preflight response is not successful`

To solve this, we are now checking if the error is a `TypeError` instead
of checking for the error's message.